feat: validate docker repository config, refactor
